requirements.txt
flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.76
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0